Reduce code repetition around get/set_register_value
