+++ CMakeLists.txt
@@ -2372,6 +2372,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -2576,6 +2577,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -4699,6 +4701,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -4903,6 +4906,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -5107,6 +5111,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -8191,6 +8196,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -8395,6 +8401,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -9560,6 +9567,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -10723,6 +10731,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -11888,6 +11897,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -13053,6 +13063,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -14218,6 +14229,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -15383,6 +15395,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -16548,6 +16561,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -17713,6 +17727,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -18878,6 +18893,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
@@ -20043,6 +20059,7 @@ if(REF_BUILD_FRAMEWORK AND CMAKE_SIZEOF_VOID_P EQUAL 8) # build-framework
 		"src/D3D12Hook.hpp"
 		"src/DInputHook.cpp"
 		"src/DInputHook.hpp"
+		"src/DisasmUtils.hpp"
 		"src/ExceptionHandler.cpp"
 		"src/ExceptionHandler.hpp"
 		"src/Genny.hpp"
